requirements.txt
python-dotenv==1.0.0
Pillow>=9.0.0 